moves request for all lifts to LiftsController
